refactor(dsl-js): Extract mapThrownError and its helpers into thrownError.ts
